LayoutTests/ietestcenter/Javascript/15.2.3.10-0-1-expected.txt

15.2.3.10-0-1 (TestCases/chapter15/15.2/15.2.3/15.2.3.10/15.2.3.10-0-1.js) Object.preventExtensions must exist as a function

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ES5Harness.preconditionPassed is true
FAIL ES5Harness.testPassed should be true (of type boolean). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
